feat: Add DashboardPage component and update routing; refactor InventoryProductPage and sidebar items
