+++ client/src/features/products/InventoryProductPage.tsx
@@ -1,6 +1,6 @@
 // COMPONENTS
-import { Loading, ProductStatsBanner, ScrollContainer, Section, SectionContainer } from "@/components";
-import { CreateByCsvBtn, CreateProductBtn, CreateProductModal, DeleteAlert, FilterBtn, FilteredStatsBanner, HealthChart, InventorySearchbar, ProductTable, ProfitChart, StockChart, UpdateStockModal } from "@/features/products/components";
+import { Loading, ScrollContainer, SectionContainer } from "@/components";
+import { CreateByCsvBtn, CreateProductBtn, CreateProductModal, DeleteAlert, FilterBtn, FilteredStatsBanner, InventorySearchbar, ProductTable, UpdateStockModal } from "@/features/products/components";
 import { ExportBtn } from "@/features/products/components/Export/ExportBtn";
 import { NukeBtn } from "@/features/products/components/ui/NukeBtn";
 import { useProducts } from "@/features/products/hooks";
@@ -10,22 +10,9 @@ import { useRef } from "react";
 
 // MAIN SECTION
 export const ProductInventorySection = () => {
-    const { isLoading, products } = useProducts();
+    const { isLoading } = useProducts();
 
     const csvInputRef = useRef<HTMLInputElement>(null);
-    const scrollToTable = () => {
-        const viewport = document.getElementById('main-content-viewport');
-        const table = document.getElementById('inventory-table');
-
-        if (viewport && table) {
-            const targetScroll = (table.getBoundingClientRect().top + viewport.scrollTop) - viewport.getBoundingClientRect().top;
-
-            viewport.scrollTo({
-                top: targetScroll - 20,
-                behavior: "smooth"
-            });
-        }
-    };
 
     if (isLoading) return <Loading label="Fetching Data..." />;
 
@@ -35,59 +22,16 @@ export const ProductInventorySection = () => {
 
             <ScrollContainer id="main-content-viewport">
 
-                {products.length > 0 && (
-                    <>
-                        {/* KPIs */}
-                        <motion.div
-                            initial={{ opacity: 0, x: -100 }}
-                            animate={{ opacity: 1, x: 0 }}
-                            transition={{ duration: 1.3, ease: "backInOut" }}
-                        >
-                            <SectionContainer>
-                                <ProductStatsBanner />
-                            </SectionContainer>
-                        </motion.div>
-
-                        <Divider />
-
-                        {/* CHARTS */}
-                        <motion.div
-                            initial={{ opacity: 0, x: 100 }}
-                            animate={{ opacity: 1, x: 0 }}
-                            transition={{ duration: 1.3, ease: "backInOut" }}
-                        >
-                            <SectionContainer size="xs">
-
-                                <Section size="lg" title="Daily Inventory Value">
-                                    <ProfitChart />
-                                </Section>
-
-                                <Section size="lg" title="Stock Distribution">
-                                    <StockChart onChartClick={scrollToTable} />
-                                </Section>
-
-                                <Section size="lg" title="Health Check">
-                                    <HealthChart onChartClick={scrollToTable} />
-                                </Section>
-                            </SectionContainer>
-                        </motion.div>
-
-                        <Divider />
-
-                        {/* FILTERED KPIs */}
-                        <motion.div
-                            initial={{ opacity: 0, x: -100 }}
-                            animate={{ opacity: 1, x: 0 }}
-                            transition={{ duration: 1.3, ease: "backInOut" }}
-                        >
-                            <SectionContainer>
-                                <FilteredStatsBanner />
-                            </SectionContainer>
-                        </motion.div>
-
-                        <Divider />
-                    </>
-                )}
+                {/* FILTERED KPIs */}
+                <motion.div
+                    initial={{ opacity: 0, x: -100 }}
+                    animate={{ opacity: 1, x: 0 }}
+                    transition={{ duration: 1.3, ease: "backInOut" }}
+                >
+                    <SectionContainer>
+                        <FilteredStatsBanner />
+                    </SectionContainer>
+                </motion.div>
 
                 {/* SEARCH, FILTER, CREATE, EXPORT */}
                 <motion.div
@@ -103,6 +47,7 @@ export const ProductInventorySection = () => {
                                 <FilterBtn />
                             </div>
 
+                            {/* CREATE , EXPORT, NUKE */}
                             <CreateProductBtn onCsvClick={() => csvInputRef.current?.click()} />
                             <ExportBtn />
                             <NukeBtn />
