+++ client/src/features/products/index.ts
@@ -1,2 +1,3 @@
+export * from './DashboardPage';
 export * from './InventoryProductPage';
 export * from './state';
