+++ server/src/main/resources/.env.example
@@ -1,0 +1,6 @@
+# .env.example
+# Copy this file to .env and fill in your values
+
+DB_URL=jdbc:postgresql://<host>:5432/<database>
+DB_USERNAME=yourusername
+DB_PASSWORD=yourpassword
