+++ server/src/main/resources/application.properties
@@ -11,6 +11,3 @@ spring.jpa.properties.hibernate.default_schema=public
 # JPA / HIBERNATE
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
-
-# SECURITY
-app.security.jwt.secret=${JWT_SECRET}
